requirements.txt
dj-database-url==0.5.0
Django==2.0.5
gunicorn==19.9.0
psycopg2==2.7.5
pytz==2018.5
whitenoise==4.0
django-taggit==0.22.2
django-cors-headers==2.1.0
django-filter==2.0.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
PyJWT==1.5.3